requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
discord.py==2.3.2
frozenlist==1.4.0
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-auth==2.22.0
google-generativeai==0.1.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
idna==3.4
multidict==6.0.4
openai==0.28.0
proto-plus==1.22.3
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
pytz-deprecation-shim==0.1.0.post0
regex==2023.8.8
requests==2.31.0
rsa==4.9
six==1.16.0
tiktoken==0.4.0
tqdm==4.66.1
tzdata==2023.3
urllib3==1.26.16
yarl==1.9.2
